refactor part 1 done
