+++ commands/samehat.js
@@ -6,7 +6,7 @@ const getAvatar = require('../utils/avatarCheck');
 module.exports = {
     async execute(message) {
         if (!message.mentions.users.size) {
-			message.channel.fetchMessages({
+			message.channel.messages.fetch({
 				limit: 1,
 				before: message.id,
 			})
@@ -37,7 +37,7 @@ async function callHat(message, otherHat) {
 	ctx.arc(127, 195, 40, 0, Math.PI * 2);
 	ctx.closePath();
 	ctx.clip();
-	const buffer = await getAvatar(message.member.user.displayAvatarURL);
+	const buffer = await getAvatar(message.member.user.displayAvatarURL({ format: 'png', dynamic: true, size: 512 }));
 	const avatar = await Canvas.loadImage(buffer);
 	ctx.drawImage(avatar, 87, 155, 80, 80);
 	ctx.restore();
@@ -45,10 +45,10 @@ async function callHat(message, otherHat) {
 	ctx.arc(370, 208, 50, 0, Math.PI * 2);
 	ctx.closePath();
 	ctx.clip();
-	const buffer2 = await getAvatar(otherHat.displayAvatarURL);
+	const buffer2 = await getAvatar(otherHat.displayAvatarURL({ format: 'png', dynamic: true, size: 512 }));
 	const avatar2 = await Canvas.loadImage(buffer2);
 	ctx.drawImage(avatar2, 320, 158, 100, 100);
-	const attachment = new Discord.Attachment(canvas.toBuffer(), 'samehat.png');
+	const attachment = new Discord.MessageAttachment(canvas.toBuffer(), 'samehat.png');
 	message.channel.send(attachment);
 }
 module.exports.info = {
