+++ commands/somad.js
@@ -1,24 +1,13 @@
 // imports
-const Canvas = require('canvas');
 const Discord = require('discord.js');
-const getAvatar = require('../utils/avatarCheck');
+const imageGeneration = require('../utils/imageGeneration');
 // exports
 module.exports = {
     async execute(message) {
-        message.channel.startTyping();
-        const canvas = Canvas.createCanvas(255, 68);
-		const ctx = canvas.getContext('2d');
-		const background = await Canvas.loadImage('./images/somad.png');
-		ctx.drawImage(background, 0, 0, canvas.width, canvas.height);
-		const buffer = await getAvatar(message.member.user.displayAvatarURL);
-		const avatar = await Canvas.loadImage(buffer);
-		ctx.drawImage(avatar, 200, 9, 50, 50);
-		const canvas2 = Canvas.createCanvas(510, 136);
-		const ctx2 = canvas2.getContext('2d');
-		ctx2.drawImage(canvas, 0, 0, canvas2.width, canvas2.height);
-		const attachment = new Discord.Attachment(canvas2.toBuffer(), 'custommad.png');
+		const avatarURL = message.member.user.displayAvatarURL({ format: 'png', dynamic: true, size: 512 });
+		const imageBuffer = await imageGeneration.soMadImage(avatarURL);
+		const attachment = new Discord.MessageAttachment(imageBuffer, 'custommad.png');
         message.channel.send(attachment);
-        message.channel.stopTyping();
 		try {
 			message.delete();
 		}
