+++ package.json
@@ -4,16 +4,16 @@
   "description": "Rewrite of BigJB21",
   "main": "index.js",
   "dependencies": {
+    "@discordjs/opus": "^0.3.2",
     "@ffprobe-installer/ffprobe": "^1.0.12",
     "bufferutil": "^4.0.1",
-    "canvas": "^2.6.0",
+    "canvas": "^2.6.1",
     "date-and-time": "^0.6.3",
-    "discord.js": "^11.5.1",
+    "discord.js": "^12.2.0",
     "fluent-ffmpeg": "^2.1.2",
     "gif-frames": "^1.0.1",
     "libsodium-wrappers": "^0.7.6",
-    "node-fetch": "^2.6.0",
-    "opusscript": "0.0.7"
+    "node-fetch": "^2.6.0"
   },
   "author": "Ramona Prower",
   "license": "ISC"
